requirements.txt
streamlit
python-docx
deep-translator
requests
